Optimized routes functions and distributed them in separate file
