+++ src/gameloop/handledata.rs
@@ -1,6 +1,10 @@
+pub mod handle_account;
+pub mod handle_action;
+pub mod handle_general;
+pub mod handle_item;
+pub mod handle_trade;
 pub mod mapper;
 pub mod router;
-pub mod routes;
 
 pub use mapper::PacketRouter;
 pub use router::{SocketID, handle_data};
